+++ R/webutils-package.R
@@ -1,8 +1,8 @@
 #'
 #' Functions for Accessing the Web
 #'
-#' Functions useful for accessing Github via the API, along with a few other
-#' functions helpful for working on the web.
+#' Functions useful for accessing Github and PyPi via the API, along with
+#' a few other functions helpful for working on the web.
 #'
 #' @name webutils-package
 #' @aliases webutils-package webutils