+++ NEWS.md
@@ -1,2 +1,5 @@
+# webutils 0.2.0 2022-05-17
+* Add functions for searching PyPi.org
+
 # webutils 0.1.0 2022-05-05
 * First release.